+++ docx4all/src/main/java/org/plutext/client/wrappedTransforms/TransformFailed.java
@@ -36,7 +36,7 @@ public class TransformFailed extends TransformAbstract {
     	super(t);
     }
 
-    public int apply(ServerFrom serverFrom)
+    public long apply(ServerFrom serverFrom)
     {
         log.warn("TransformFailed not fully implemented!");
